refactor sign up page to include success state
